CHG: Correction in the insertion of the partner ID of the partner who creates the guest

+++ club/src/main/java/app/service/Service.java
@@ -1,24 +1,25 @@
 package app.service;
 
-import java.sql.Date;
 import java.sql.SQLException;
 
 import app.dao.*;
 import app.dao.interfaces.*;
+import app.dto.GuestDto;
 import app.dto.PartnerDto;
 import app.dto.PersonDto;
 import app.service.interfaces.AdminService;
 import app.service.interfaces.LoginService;
 import app.service.interfaces.PartnerService;
-import app.service.interfaces.GuestService;
 import app.dto.UserDto;
+import app.helpers.Helper;
+import app.model.Partner;
 
 
 public class Service implements AdminService, LoginService , PartnerService{
-    private PersonDao personDao;
-    private UserDao userDao;
-    private PartnerDao partnerDao;
-    private GuestDao guestDao;
+    private final PersonDao personDao;
+    private final UserDao userDao;
+    private final PartnerDao partnerDao;
+    private final  GuestDao guestDao;
     private DetailInvoiceDao detailInvoiceDao;
     private InvoiceDao invoiceDao;
     
@@ -28,6 +29,7 @@ public class Service implements AdminService, LoginService , PartnerService{
         this.personDao = new PersonDaoImplementation();
         this.userDao = new UserDaoImplementation();
         this.partnerDao = new PartnerDaoImplementation();
+        this.guestDao = new GuestDaoImplementation();
     }
     
     @Override
@@ -36,21 +38,36 @@ public class Service implements AdminService, LoginService , PartnerService{
     }
     
     @Override
-    public void createGuest(UserDto userDto) throws Exception{
-        this.createUser(userDto);
+    public void createGuest(GuestDto guestDto) throws Exception{
+        this.createGuestInDb(guestDto);
+    }
+    
+    @Override
+    public void activateGuest(GuestDto guestDto) throws Exception{
+        this.activateGuestInDb(guestDto);
+    }
+    
+    @Override
+    public void inactivateGuest(GuestDto guestDto) throws Exception{
+        this.inactivateGuestInDb(guestDto);
     }
     
     @Override
     public void login(UserDto userDto) throws Exception {
         UserDto validateDto = userDao.findByUserName(userDto);
+        
         if (validateDto == null) {
             throw new Exception("no existe este usuario registrado");
         }
+        
         if (!userDto.getPassword().equals(validateDto.getPassword())) {
             throw new Exception("usuario o contraseña incorrecto");
         }
+        
         userDto.setRole(validateDto.getRole());
-        user = validateDto;
+        userDto.setPersonId(validateDto.getPersonId());
+        userDto.setId(validateDto.getId());
+        user = userDto;
     }
         
     @Override
@@ -63,17 +80,23 @@ public class Service implements AdminService, LoginService , PartnerService{
         if (this.personDao.existsByDocument(personDto)) {
             throw new Exception("ya existe una persona con ese documento");
         }
+        
         this.personDao.createPerson(personDto);
+    }
+    public void requestToUnsubscribe(PartnerDto partnerDto) throws Exception{
+        
     }
     
     private void createUser(UserDto userDto) throws Exception {
         this.createPerson(userDto.getPersonId());
         PersonDto personDto = personDao.findByDocument(userDto.getPersonId());
         userDto.setPersonId(personDto);
+        
         if (this.userDao.existsByUserName(userDto)) {
             this.personDao.deletePerson(userDto.getPersonId());
             throw new Exception("ya existe un usuario con ese user name");
         }
+        
         try {
             this.userDao.createUser(userDto);
         } catch (SQLException e) {
@@ -81,14 +104,60 @@ public class Service implements AdminService, LoginService , PartnerService{
         }
     }
     
-    private void createPartnerInDb(PartnerDto partnerDto) throws Exception{
+    
+    private void createPartnerInDb(PartnerDto partnerDto) throws Exception {
         this.createUser(partnerDto.getUserId());
         UserDto userDto = userDao.findByUserName(partnerDto.getUserId());
+        PersonDto personDto = personDao.findByDocument(partnerDto.getUserId().getPersonId());
         partnerDto.setUserId(userDto);
+        
+        if (partnerDto.getType().equalsIgnoreCase("VIP") && this.partnerDao.countPartnersVip() >= 5 ) {
+            this.userDao.deleteUser(userDto);
+            this.personDao.deletePerson(personDto);
+            throw new Exception("Ya existen 5 socios VIP");
+        }
+
         try {
             this.partnerDao.createPartner(partnerDto);
-        } catch(Exception e){
-            this.userDao.deleteUser(userDto);
+        } catch (SQLException e) {
+            System.out.println("Ocurrio un error: " + e.getMessage());
         }
+    }
+
+    
+    private void createGuestInDb(GuestDto guestDto) throws Exception{
+        this.createUser(guestDto.getUserId());
+        
+        UserDto userDto = userDao.findByUserName(guestDto.getUserId());
+        PersonDto personDto = personDao.findByDocument(guestDto.getUserId().getPersonId());
+        guestDto.setUserId(userDto);
+        
+        try {
+            this.guestDao.createGuest(guestDto);
+        } catch(SQLException e){
+            System.out.println("Ocurrio un error: " + e.getMessage());
+            this.userDao.deleteUser(guestDto.getUserId());
+            this.personDao.deletePerson(personDto);
+        }    
+    }
+    
+    private void activateGuestInDb(GuestDto guestDto) throws Exception{
+       
+    }
+    
+    public void inactivateGuestInDb(GuestDto guestDto) throws Exception{
+     
+    }
+
+    @Override
+    public PartnerDto getSessionPartner() throws Exception {
+        if (user == null) {
+            throw new Exception("No hay usuario en sesión.");
+        }
+        PartnerDto partnerDto = partnerDao.findByUserId(user);
+        if (partnerDto == null) {
+            throw new Exception("No se encontro el socio asociado al usuario en sesion.");
+        }
+        return partnerDto;
     }
 }
